chromeos-base/chrome-icu: sync build dependencies
We have an issue where the new icu ebuild wasn't correctly
pulling in build dependencies as well as specifying one
which chrome's ebuild no longer uses. We've modified an
existing uprev with the changes hoping it would go through
but it unfortunately appears to be a bad version of chrome.

So we'll make the changes here instead and expect future
PUpr runs to pull these modifications in.

BUG=chromium:1062006
TEST=identical ebuild builds on current uprev

Change-Id: Idef39431668ed1a506774dfc4c244e47a6c2a9f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2114080
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>