usbguard: Use usb_bouncer generated policy.
This uses usb_bouncer to generate the usbguard policy. This will
hopefully address timeouts that occurred in the previous policy
generation. Additionally the generated policy was changed to be closer
to the original policy to have a smoother transition to the new policy
generator.

BUG=chromium:992209
TEST=tast run device security.USBGuard

Change-Id: I961076ea499bc272108ae1880dd144cac151e6ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2120249
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>