FROMLIST: sys-apps/util-linux: Add nosymfollow flag
In the 5.4 kernel, the new MNT_NOSYMFOLLOW mount syscall flag replaces
the existing nosymfollow option that was previously implemented through
the chromiumos LSM.  This adds support to the mount command (as
-onosymfollow).  Since this is the same name that we had already used,
no callers need to be changed.

We don't want to use the new flag on older kernels, so the patch is only
applied on 5.4.  Rather than specifically checking for 5.4, this
excludes all the older kernel versions so that newer kernels will
automatically get the patch.

The patch and MNT_NOSYMFOLLOW flag have both been sent upstream.  We
should be able to revert back to portage-stable once everything has made
it into an upstream release.

BUG=b:152074038
TEST=mount -onosymfollow and try to follow a symlink on 5.4 and 4.19.

Change-Id: I23ed478fc6a67a0653d561882a18220aad072ded
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2145680
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>