rmtfs: Uprev to 20200203
Uprev rmtfs to the latest commit currently available on Github. I've
been carrying around a revert of this commit in private modem kernel
sandboxes for awhile. This uprev will allow me to drop the revert of:

4107102d60791 remoteproc: qcom_q6v5: don't auto boot remote processor

With the latest rmtfs, the userspace daemon does a little more. It's now
capable of starting and stopping the remote processor when the daemon
itself starts and stops. As such, it needs to write a sysfs file to do
that, hence the added bind mount on /sys and the chown in the
pre-start script. It also spawns a new thread to do the start/stop work,
hence all the additional syscalls in the seccomp filter.

BUG=b:74020136
TEST=Boot with new kernel and new rmtfs, see modem run

Change-Id: I6c51ce9ef69eab084b4db5f27a2560cfc5c57110
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2134590
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>