libtextclassifier: remove model installation.
This CL reverts the two previous CLs that try to install the
flatbuffer model in libtextclassifier.ebuild. Although, logically, the
model file should be installed in this this ebuild, because the
unit-test framework of ml-service requires copying the model file, for
simplicity, we'd better install the model file in ml.ebuild to avoid
mentioning the model file twice (once here, once in ml.ebuild), which
would cause inconvenience in uprev.

BUG=chromium:1020396
TEST=emerge-eve libtextclassifier works.

Change-Id: Ie2ee230278a3fa29c6093aa02609c2b2ac9a1eca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2135679
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>