bluez: Use make install
Using our own install script in ebuild turns out to be hard after
bluez-next. This patch moves it back to using make install. To solve the
original intention to selectively install files, it can be done with
improving the configure script upstream.

This is a semi-revert of CL:2039993.

BUG=chromium:1019578,b:150851302
TEST=Build and check the installed files.

Change-Id: I2389f1efe3d407daf1476db4656582aaad6cc6e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2138098
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>