camera-intel-ipu6: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes it clearer
that the `switch` above catches all interesting cases. If someone does
add a case that doesn't `return` down the line, they'll get a -Werror'ed
warning about lacking a `return` with a valid value.

BUG=chromium:1066998
TEST=CQ

Change-Id: Ifea0cc373303138cf7a0c44730d26884bbf2a0ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2143385
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Reviewed-by: Liang L Yang <liang.l.yang@intel.com>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>