app-emulation/lxd: patch libco with inline asm
As a short-term workaround for crbug.com/1059078 we marked the co_switch
function as needing to be generated with arm code instead of thumb.
However, this still left undefined behavior in place which could bite us
again in the future. These patches switch the co_swap implementation
from a data array to a function implemented in assembly.

Fixes were submitted to canonical's libco tree in
https://github.com/canonical/libco/pull/13

BUG=chromium:1059078
TEST=start lxd on tatl/tael and hacked-together tael64

Change-Id: I46fb5fd459baca71188819f9ab2800234cc55f5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2144506
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>