chromium-source: add checking for git-cache locks for short term fix
The commit checks /tmp/git-cache for *.lock files and does a few things.
If it finds a lock file older than 30 minutes it assumes that it is
left over and simply adds the --ignore_locks flag to gclient sync. If it
finds locks in the age of <30 minutes it then sleeps 15 minutes and
checks them again. If the locks are still there, it still adds the
--ignore_locks flag. This is meant as a temporary measure while we root
cause the reason for the left-over locks.

BUG=chromium:1069842
TEST=emerge-eve chrome-icu

Change-Id: I550b94318515d9669ad1e5ade63cc1ab085b12d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2151060
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: David Burger <dburger@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>
Tested-by: George Engelbrecht <engeg@google.com>