media-libs/mesa-amd: Pass through WC information for imported buffers.
This ensures that radeonsi keeps track of uncached/cached state of
imported buffers. This way it can make a sensible selection on
whether to use a staging texture for memory mapping.

Without a staging texture, reading from uncached memory is very slow
and this is hurting is with video decoding as vaGetImage does
an image read from the CPU.

BUG=b:152378755
TEST=Run video.DecodeAccel.h264_resolution_switch on grunt.

Change-Id: If3686e0bf4371d0bfaa9b47f57eb560f496d31af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2153211
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Shirish S <shirish.s@amd.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>
Commit-Queue: Bas Nieuwenhuizen <basni@chromium.org>