linux-firmware: Bind kernel and iwlwifi versions
Instead of installing all versions of the wifi firmware for a particular
chipset, specify exactly which file should be installed depending on
which kernel the board is running. For "mature" boards the firmware
version is typically the same across kernel versions, but for more
recent chipsets and during bringup it is relatively common to have
different firmware versions for different kernels.

This reduces disk usage of the rootfs since we only install the
necessary firmware file.

Original discussion at http://shortn/_8jBvFFO3xX

BUG=b:149998064
TEST=emerge-hatch linux-firmware
TEST=emerge-volteer linux-firmware
TEST=USE="-kernel-4_19" emerge-hatch linux-firmware -> warnings about
kernel not being set and all iwlwifi-QuZ-* files are installed.

Change-Id: I44b0bca4073366cbf539fca189ad414fe76433be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2153883
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>