hwsec-test-utils: initial commit
This CL creates a new project "hwsec-test-utils" that only installed in
test images.

It is part of preparation for the upcoming test utilities used in
integration test.

There are some dependencies and placeholders that are necessary in the
near future though it doesn't make 100% sense to be in the initial commit.

BUG=b:154672168
TEST=emerge-chell hwsec-test-utils

Cq-Depend: chromium:2160295
Change-Id: I896300e4c21e0ee2b9cacaf63c1c02c0da561f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2160371
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>