volteer: Add trondo coreboot configuration
Create a new coreboot configuration for the trondo variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.0.1).

BUG=b:154678884
TEST=FW_NAME=trondo emerge-volteer coreboot chromeos-bootimage
Ensure that image-trondo.*.bin are created

Cq-Depend: chrome-internal:2932321
Change-Id: I6a6a5d19e869b533ad399b1e4afcda4841069cbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2166853
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>