afdo: Enable AFDO profile for kernel 4.19
We have been using kernel AFDO profile on banon for a couple of
months now. Experiments on beta branch show it reduces the CPU
loads a lot compared with other boards in the same family. So
we want to enable using the profile for all boards using kernel
4.19.

BUG=b:154543384
TEST=Will use tryjobs and CQ

Change-Id: I6f7605cebf0555135523ae781a8f4f38650a419b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2171057
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Tiancong Wang <tcwang@google.com>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>