chrome-binary-tests: Add a note about Chrome's //:chromiumos_preflight.
A while ago, a target in this list failed to compile and broke some
builds. Since it wasn't in Chrome's chromiumos_preflight target,
Chrome's CQ didn't compile it, and we didn't catch the build breakage.

Add a comment to the top of the list so hopefully folks will know to
keep chromiumos_preflight in sync.

BUG=none
TEST=none

Change-Id: Ib77eeca30e9c6cd6b716927c29915de2f39da46f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2181670
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>