Add arc-toolchain-r ebuild package
This is the first commit for the new Android RVC arc toolchain package,
it has been manually uprevved and uploaded following the instructions on
the gather.sh script (carried over from the arc-toolchain-master
package). No boards are using this yet so it hasn't been tested but it
provides a baseline for future uprevs.

BUG=b:152934175
TEST=None

Change-Id: I8cc0e02dcf7f96be8d6b88f01bb528162ce9f193
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2190430
Commit-Queue: Morg <morg@chromium.org>
Tested-by: Morg <morg@chromium.org>
Reviewed-by: Richard Fung <richardfung@google.com>
Auto-Submit: Morg <morg@chromium.org>