quickoffice: use CrOS-side upstart script
Instead of using 2 upstart scripts that are coming from the Quickoffice
tree and make it difficult to work on them, handle the upstart script
that mounts/unmounts quickoffice at runtime in the Chrome OS source
tree.

Also, instead of 2 upstart scripts (1 for mounting, 1 for unmounting),
use a single upstart script that does both the mounting and the
unmounting.
Do not depend on the "ui" job, only mount once on boot and unmount when
shutting down the device. That will also avoid race conditions when
upstart tries to mount the quickoffice squashfs before it has been
unmounted.

BUG=chromium:1044720
TEST=CQ dry-run, flash the hatch image artifact on a device. Verify that
the quickoffice squashfs is mounted automatically on boot and
Quickoffice still works.

Change-Id: I47186b703e8397d6292c1f765a157898e56217dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2191249
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>