chromeos-base/nnapi: Add patches to support logging.
To get android-base/logging.h and is supporting implementation files to
work on ChromeOS we need to make some small modifications. We do this in
a patch file that we apply to the source at prepare phase. These changes
are:

- Change liblog_symbols.cpp to not require liblog to be statically
linked on non ANDROID platforms. This allows base logging to fallback to
stderr logging.
- std::regex was crashing on a regex string in the test, so we have
commented out this part of the testing.
- Added __android_log_is_loggable to the list of functions that can be
stubbed out.
- The DCHECK() message only prints "Check Failed:", changed the string
the test was searching for.

BUG=b:156714501
TEST=FEATURES="test" emerge-octopus chromeos-base/nnapi

Change-Id: Id8677fcc0260a05d3e1f4198952de28cb5cd7ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2210282
Reviewed-by: Alan Green <avg@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>