chromeos-bootimage: Allow coreboot ebuild when zork use flag is set
We are transitioning from coreboot-zork to coreboot. This change will
allow the coreboot ebuild to be used when the zork use flag is set. This
will allow building chromeos-boot image with either coreboot-zork or
coreboot.

coreboot-zork is still the default package that will be installed if
either coreboot or coreboot-zork are not installed.

BUG=b:147042464
TEST=Tried various emerge commands:
$ emerge-zork coreboot chromeos-bootimage
$ emerge-zork coreboot-zork chromeos-bootimage
$ emerge-zork --unmerge coreboot-zork && emerge-zork chromeos-bootimage

Change-Id: Id9983c3e155cfad85464080058b5fda22d9a15ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2225088
Tested-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Raul E Rangel <rrangel@chromium.org>
Auto-Submit: Raul E Rangel <rrangel@chromium.org>