json-glib: Disable pointer-to-int-cast warning harder
Disable -Wpointer-to-int-cast warning in json-glib, this time by
yanking the line out of the meson.build file entirely.  With the uprev
to meson 0.54, the order of externally supplied CFLAGS and what the
build system comes up with has been swapped.  This means that the
-Wno-void-pointer-to-int-cast option gets added before the option to
turn it on. Let's instead prevent the option from getting added in the
first place.

BUG=chromium:1091506
TEST=json-glib builds with uprev'ed meson

Change-Id: Ib338160e130e0de80e4bdbef074dd96579c26ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2231704
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Rob Clark <robdclark@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>