lorgnette: Depend on sane-airscan
Now that we can directly pass device URLs to sane-airscan, it's useful
to have it on the board for testing.  We have autodiscovery disabled in
sane-airscan, so pulling this in won't make devices start otherwise
showing up before we're prepared to expose them to users.

BUG=b:153542563
TEST=build_packages

Change-Id: I890c087f1a5a6d7b3bcc0b119e1e5eb490cf89d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2233185
Reviewed-by: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>