Make biod's dbus proxy library its own ebuild
This is similar to other *-client ebuilds, but includes more
functionality since working with biod requires extra logic. This
ebuild is supposed to be built and installed on all boards, while
the biod ebuild is built only when the "biod" USE flag is set.

BUG=chromium:1090823
TEST=FEATURES=test emerge-nami biod
	==> Observe that biod_proxy is built (and tested) first.

Cq-Depend: chromium:2229738
Change-Id: Ie3f0e734d309fcbd28fdc58b7fe385b2e8428f28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2233845
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>