media-libs/mesa: Add dependency on x11-libs/libxshmfence
If I setup the board "tatl" and just tried to emerge mesa I'd end up at:

  mesa-20.0.7-r0: Run-time dependency xshmfence found: NO (tried pkgconfig and cmake)
  mesa-20.0.7-r0:
  mesa-20.0.7-r0: meson.build:1623:6: ERROR: Dependency "xshmfence" not found, tried pkgconfig and cmake

We're missing a dependency.  Add it.  The tatl board is the cros based
guest environment we use for crostini, which needs to build a mesa
virgl driver that supports X.

BUG=None
TEST=emerge-tatl mesa

Change-Id: Iff3aa978e051e1290b44c8f909b6862bf8c275b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2235925
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>