autotest-tests-audio: Add audio_AudioInputGain
Test audio_AudioInputGain depends on telemetry. It requires the path to
telemetry source to exist in order to build. Copy the telemetry source
to a temporary directory that is writable, so that file removals in
Telemetry source can be performed properly.

BUG=b:157443945
TEST=emerge-soraka autotest
TEST=cros_workon-soraka start autotest-tests-audio
TEST=emerge-soraka autotest-tests-audio

Change-Id: I91a0fd541cba529de9e44f3f54ebc1edad4de23a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2237209
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>