dedede: Add drawcia coreboot configuration
Create a new coreboot configuration for the drawcia variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.0.1).
manually change spi descriptor/me name to drawcia.

BUG=b:158540280
TEST=FW_NAME=drawcia emerge-dedede coreboot chromeos-bootimage
Ensure that image-drawcia.*.bin are created

Cq-Depend: chromium:2242676, chrome-internal:3093955
Change-Id: I683e8e57a47ed59852cc58722b3e906aa6bbd852
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2237547
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>