mesa: Turn off shader cache using meson flag
Doug says: "When emerging mesa a quiet little note flew by.  No
failures.  No capital letters.  No asterisks.  No explanation points.
...but I still noticed it and it was so cute and little I felt sorry
for it, so I decided to fix it."

The error was:

  please use append-cppflags for preprocessor flags

but we now have a proper meson option for it, so lets use that.

BUG=none
TEST=The cute little note is gone and I miss it already.

Change-Id: I78f336d0c22ff67cdcd63554cc60fdce509e61c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2235926
Tested-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>