net-misc/rmtfs: Don't put rmtfs in a new PID namespace
Putting rmtfs in a PID namespace seems to affect its ability to
gracefully shut down the modem and stop itself. This affects reboot
time.

I'm not fully sure why the PID namespaces cause this issue.
Part of the story is that when using PID namespaces, minijail sticks
around to act as PID 1 in that new namespace. With 'expect fork',
upstart is actually tracking minijail, not rmtfs. When stopping, it
kills minijail, resulting in a SIGKILL going to rmtfs.

The part that's not understood is that when changing this to 'expect
daemon', so that the rmtfs process itself is tracked, 'stop rmtfs', and
reportedly subsequent 'start rmtfs' attempts just hang.

BUG=b:158029675 b:158059671
TEST=stop rmtfs

Change-Id: I1771f49f8bab32f5bf22f78b01e7caab10e65224
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2283842
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>