unibuild: remove cros_config_host_local
This function is used by the cros-firmware eclass to dynamically
generate SRC_URI based on the model.yaml file.  It does not work, for
a few reasons:

1. Execution of external commands (cros_config_host) is not allowed
   when the ebuild is being sourced.
2. Nobody actually puts the model.yaml file at the hard-coded path
   here.
3. The logic to trim the filename from ${EBUILD} is wrong...

Instead, the current practice is to generate files/srcuris next to the
ebuild when updating the firmware.  So just delete this broken
logic...

BUG=chromium:927917
TEST=emerge-nocturne chromeos-firmware-nocturne

Change-Id: If91728a830bcbee1c9e5339aec2107f0eb0206f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2248664
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>