Add a perfetto ebuild
Perfetto is a tracing daemon that traces Android and Linux. More details
are at perfetto.dev. It is embedded into Chrome; however we would like
it to be able to do kernel tracing. For this we would like the trace
daemon (traced) available outside of Chrome. Further, we would like
traces to be collectible on the command line. This is the first step
toward getting Perfetto on ChromeOS working. More details at:
go/cros-perfetto.

This change adds an ebuild for building standalone perfetto build. This
allows running perfetto on the command line and collecting kernel
traces.

Co-authored-by: Joel Fernandes <joelaf@google.com>
BUG=b:147789115
TEST=emerge-${BOARD} perfetto for hatch and kukui

Change-Id: Ic355dc3b725e3c91718a76ab7bdc339ad92f8968
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2251681
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>