volteer: Add volteer2 coreboot configuration
Create a new coreboot configuration for the volteer2 variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:159135047
TEST=FW_NAME=volteer2 emerge-volteer coreboot chromeos-bootimage
Ensure that image-volteer2.*.bin are created

Change-Id: I17389ce269ee5f2eb8fbda42350b69be8f460676
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2255286
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: YH Lin <yueherngl@chromium.org>
Commit-Queue: YH Lin <yueherngl@chromium.org>