volteer: Add delbin coreboot configuration
Create a new coreboot configuration for the delbin variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:158797761
TEST=FW_NAME=delbin emerge-volteer coreboot chromeos-bootimage
Ensure that image-delbin.*.bin are created

Change-Id: I60e0808c586d8a994a877bc02f664452a34a195c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2262097
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>