Enable distributed ThinLTO for arm and aarch64 boards
Distributed ThinLTO allows builds to complete in less wall time.
Support has been added to use Goma to distribute code generation
for ThinLTO. This is currently enabled for several boards,
and we are in the process of enabling it for more. This
change enables distributed ThinLTO for all boards using arm or
aarch64 CPUs. Previous changes have already enabled distributed
ThinLTO for most of these boards on a per-board basis; those
changes can be removed once this change is in.

BUG=chromium:877722
TEST=emerge-${BOARD} chromeos-chrome

Change-Id: I253a633112b1e6a576a345db2e690297efb852c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2267368
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>
Commit-Queue: Bob Haarman <inglorion@chromium.org>