chromeos: disable pygobject[cairo]
We appear to enable this flag, but pygobject:2 doesn't actually set a
pycairo dependency, and pygobject:2 doesn't actually accept the
--with-cairo configure flag, so it doesn't really do anything.

On the other hand, pygobject:3 enables this flag by default, so we'd
need to explicitly disable it if we don't want it.

Skip out on cairo, so we don't have to teach pycairo how to
cross-compile correctly.

BUG=chromium:1099508
TEST=build

Change-Id: If4bb2e394ac560e348ccb91737106fb199a1b416
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2269257
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>