libva-intel-media-driver: Fix VP9 Decoder to avoid GPU HANG (patch 1)
This patch fixes the some of the gpu hangs observed with vp9 decode
on GEN9 platform due to the wrong handling of the
SegmentIDStreamIn & UsePrevInFindMvReferences

Upstream PR: https://github.com/intel/media-driver/pull/957

a) The current code tries to reset SegmentIDStreamIn if there is
a resolution change on inter-frames. But due to the mishandling,
the code was eventually setting SegmentIDStreamIn to true for all
non-scaling use cases. The patch ensures that the reset happens
only for res-change on P-frames irrespective of other conditions.

b) The current implementation sets the SegmentIDStreamIn for any
decode where there is no Segmentation Enabled. We rather follow
strict constraints similar to the legacy i965 implementation.

c) Ensure the correct SegmentIDStreamIn for key-frames, intra-only
frames and error-resilient mode.
d) Reset UsePrevInFindMvReferences to zero if last picture has a different size,
Current picture is error-resilient mode, Last picture was intra_only or keyframe,
Last picture was not a displayed picture.

Thanks to Wang, Yanyan.wang@intel.com for the review
and Li, Steven(steven.li@intel.com) for the suggestion
on keeping the scaling logic outside of the inter-frame code block.

BUG=b:152554969
TEST=Play the vp9 stream downloaded from the partner bug 152554969
(https://drive.google.com/open?id=1VxyD5SVj7rUkTsDPNjcAuFQ7nwae_WNN)
a couple of times and ensure that there is no GPU hang.

Change-Id: I793b1895265329e2348d1b919378416a7bc4cd9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2229422
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>