eclass: use skylark for toolchain configuration
In preparation for upgrading Bazel, we need to define the toolchain in
Skylark and not in a config proto. This also sheds a lot of features
that were defined but not used in our configuration.

This version of the eclass is compatible with Bazel 0.24.1 (current) and
Bazel 3.2.0 (future). This is one less moving part for the main TFLite
uprev.

BUG=chromium:1102795,b:159987197
TEST=FEATURES=test emerge-{betty,kukui} chromeos-base/drivefs \
                                        sci-libs/tensorflow

Change-Id: Iac15c3a55beaf6d646f479e50e759b47322e0851
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2275763
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Auto-Submit: Jim Pollock <jmpollock@chromium.org>