libva-intel-driver: Disable X11 and Wayland.
This CL explicitly disables X11 and Wayland for libva-intel-driver
because we never use VA-API over X11 or Wayland on Chrome OS. This
involves removing the corresponding IUSE flags.

This is a cleanup to make it possible for the libva uprev to also not
involve IUSE flags that we don't care about.

Note that it is not necessary to pass --enable-drm: this is not a
supported flag [1].

[1] https://github.com/intel/intel-vaapi-driver/blob/2.3.0/configure.ac#L50-L68

BUG=b:159825228
TEST=./build_packages --board=eve

Cq-Depend: chromium:2285394
Change-Id: Ie8a29dac90aae20bf24053151771ab3d3686f5ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278031
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>