iHD: Re-apply "Encoder-VP8-GEN9-GEN10-GEN11-Ensure-forced_lf_adjust."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:1788518. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's upstream starting on
intel-media-20.2.pre1 [1].

Note that I regenerated the .patch by cherry picking from upstream
instead of using the file previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/commit/2452327d78d2513595274c0aaa93d57a1cc83c8c

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:1788518 for functional testing (I did not repeat it)

Cq-Depend: chromium:2269975
Change-Id: I6e219f41b6dca0a91acb0e9af410d83b6c6a5149
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278393
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>