iHD: Re-apply "Disable-IPC-usage."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2012591. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's a Chrome-specific workaround. There's
a bug upstream [1] to make the naming more meaningful, but it's still
open.

[1] https://github.com/intel/media-driver/issues/854

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver

Cq-Depend: chromium:2278393
Change-Id: I73232ba75398b5bfa694e8a1da76c04e3cc00e7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278035
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>