iHD: Re-apply "VP-Explicitly-initialize-maxSrcRect-of-VphalRenderer."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2173244. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's upstream starting on
intel-media-20.2.pre2 [1].

Note that I regenerated the .patch by cherry picking from upstream
instead of using the file previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/commit/07a818120af884fefb6a0a4e5894a150aa5447cb

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:2173244 for functional testing (I did not repeat it)

Cq-Depend: chromium:2278035
Change-Id: I288489b572a6cb790fed428b5985bf4d41945e3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278036
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>