iHD: Re-apply "VP-fix-crash-in-vp8-playback."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2137172. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's upstream starting on
intel-media-20.2.pre3 [1].

Note that I regenerated the .patch by cherry picking from upstream
instead of using the file previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/commit/26b386f708ada4e914789e1b1855c5d4135bdd81

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:2137172 for functional testing (I did not repeat it)

Cq-Depend: chromium:2278036
Change-Id: Ic789eb04456c7c98ff89d334031ee5eac0d8b227
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278037
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>