iHD: Re-apply "VP-Fix-memory-corruption."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2174833. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's upstream starting on
intel-media-20.2.pre1 [1].

Note that I regenerated the .patch by cherry picking from upstream
instead of using the file previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/commit/3783175d95a4d27cc3be6d69e144b5f7d521cbbf

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:2174833 for functional testing (I did not repeat it)

Cq-Depend: chromium:2278037
Change-Id: I34e0c0cf251d317eb74e1efb08ca7ea3e9536796
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278395
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>