iHD: Re-apply "Decoder-VP9-GEN-9-10-11-12-Fix-the-GPU-hang-due-to-w."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2229422. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch is needed because it's upstream starting on
intel-media-20.2.pre4 [1].

Note that I regenerated the .patch by cherry picking from upstream
instead of using the file previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/commit/1aa3e6995ccfac86b4ec68265591545302435038

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:2229422 for functional testing (I did not repeat it)

Cq-Depend: chromium:2278395
Change-Id: I8f331ac6545edceabe664db44551c432f2fddbf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278614
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>