iHD: Re-apply "Decoder-VP9-GEN9-Disable-HPR-VP9-mode-switch-to-avoi."
This CL re-applies the aforementioned patch. That patch was originally
introduced by CL:2270911. It was removed in CL:2269975 with the
intention to re-apply it in its own CL to allow for bisectability.

The patch hasn't landed upstream, but it looks like merging it is
imminent [1].

Note that I cherry picked the patch from [2] instead of using the one
that was previously checked into Chrome OS.

[1] https://github.com/intel/media-driver/pull/975
[2] https://github.com/sreerenjb/media-driver/tree/gen9-vp9-gpu-hang-fix

BUG=b:159825228
TEST=emerge-eve x11-libs/libva-intel-media-driver
TEST=See CL:2270911 for functional testing (I did not repeat it)

Cq-Depend: chromium:2278614
Change-Id: I926237cfdab60bd6da256b1898eebfe8afb996ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278617
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>