iHD: Re-enable the CMRT library.
This CL allows building the iHD driver with the CMRT library as long as
the ihd_cmrtlib USE flag is set. This was originally done in CL:2143233.
It was disabled due to CL:2269975 with the intention of re-enabling it
in its own CL for bisectability.

BUG=b:159825228
TEST=./build_packages --board=dedede with USE=ihd_cmrtlib generates
     /usr/lib64/libigfxcmrt.so

Cq-Depend: chromium:2278617
Change-Id: Ibd49f953e0fca4b66c663f5316f30107504f17be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2278618
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>