chromeos-base/ec-utils: remove $BOARD selection, build with BOARD=host
This ebuild has some complicated $BOARD selection logic that gets in
the way when removing a board from a firmware branch.  Additionally,
when we're going thru and migrating pre-unibuild boards to unibuild, a
lot of the boards have been deleted from ToT and we won't be able to
find an appropriate $BOARD this way (e.g., sand).

It should be sufficient to just build with BOARD=host and skip the
BOARD selection logic... looking at the commit history, it looks like
this was added to support some (now non-existent) board-specific
features for Link back in the days.

BUG=chromium:1035536,chromium:1099753
TEST="emerge-sand ec-utils" succeeds with unibuild conversion enabled

Change-Id: Ia93c5b1ad93d8a22bd471c42e1d7194b91356fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2283999
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>