chromeos-base: eliminate syslog overflow by increasing queue size
This CL should reduce the number of failure to forward logs to
rsyslogd by increasing net.unix.max_dgram_qlen.

This changes the value of the maximum queue size for dgram unix
sockets. This defines the maximum lines of queue for syslog.
The logs are easy to be overflown when the system is high load
and receives some of multiple-line logs, since logs are sent
line by line in the syslog IPC.

This increases number to 512. The default value was 10, and the
value was increased to 60 by https://crrev.com/c/627966.

See the bug for the reason of the new value 512.

BUG=chromium:1090580
TEST=Stopped rsyslogd, sent 100-line log, resumed rsyslogd and see
     no dropped logs.

Change-Id: Ice600371b687a9c88f049650ad83397f534387ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2262092
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Tested-by: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>