mesa-llvmpipe: Move Wayland deps to RDEPEND.
While trying to uprev libva and company without Wayland support, one of
the VMs on the CQ was failing because TestImportedSymbolsAreAvailable
was complaining with an error such as:

File dir-ROOT-A/usr/lib64/libEGL.so.1.0.0 contains unsatisfied symbols:
{b'wl_display_disconnect' ...

I did not explore this fully, but what I believe was happening is that
on this particular VM, mesa was compiled with Wayland support. However,
in the final image, the wl_* symbols are not available because mesa was
'expecting' some other package to cause dev-libs/wayland to be
installed. ToT libva has dev-libs/wayland in its RDEPEND, so this is
probably the package that was installing Wayland in this VM. However,
since I'm going to force libva to not use Wayland, mesa needs to express
its own dependency on it.

Therefore, this CL moves Wayland dependencies from DEPEND to RDEPEND and
makes it look more like upstream [1].

[1] https://gitweb.gentoo.org/repo/gentoo.git/tree/media-libs/mesa/mesa-20.1.2.ebuild?id=dad4c500e21ba74856742329a05cc050dc5489be#n111

BUG=b:159825228
TEST=None

Cq-Depend: chromium:2278618
Change-Id: Ibc960119276f0db387df26b2de92551cfa3bfe70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2285394
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>