alsaucm: Fix ending with quotes commands
If we use 'alsaucm -n -b -' with 'get "CapturePCM/Internal Mic"' then
the alsaucm will report error and stop immediately. The reason is that
the parse_line in usecase.c appends an empty argument if a command ends
with quotes.

This CL adds a patch to fix the parse_line function in usecase.c.

BUG=None
TEST=(1) emerge-${BOARD} alsa-utils
     (2) printf "\"h\"\n\"h\"\n" > /tmp/test
     (3) /build/${BOARD}/usr/bin/alsaucm -n -b /tmp/test
     (4) Do (1)-(3) without the patch
     (5) Check that the output only contains help once
     (6) Do (1)-(3) with the patch
     (7) Check that the output contains help twice

Change-Id: Ic41bc9a53041ce59746b34480938565046e5199d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2286969
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>