compiler_wrapper: remove -grecord-gcc-switches
Port over compiler_wrapper change crrev.com/c/2290624 into CrOS

Description of that change:

-grecord-gcc-switches causes the compiler and compiler flags to be
stored in the produced debug information. This causes Goma builds to
be non-deterministic, because the compiler path differs depending
on whether the compiler was run locally or on Goma. This, in turn,
causes cache pollution and needless work. Since we are not in
fact using the information stored by -grecord-gcc-switches, this
change disables it. We can re-enable it when we want to start
using the information, provided we come up with another way to
avoid the issues described earlier.

BUG=chromium:1103065
TEST=Tryjobs, see that flag is no longer there

Change-Id: I3cb976fa8085e45a7c0ba25000441396b237316e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2295722
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Jian Cai <jiancai@google.com>
Commit-Queue: Bob Haarman <inglorion@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>